requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
boto3==1.34.27
requests==2.31.0
werkzeug==3.0.1
cryptography==41.0.7
Pillow==10.1.0
